UI Restored and returned Only Retrofit

+++ app/src/main/java/com/miracozkan/tvseries/datalayer/model/SeriesLastEpisode.kt
@@ -21,40 +21,40 @@ import kotlinx.android.parcel.Parcelize
 @Parcelize
 data class SeriesLastEpisode(
 
-    @PrimaryKey(autoGenerate = true)
+        @PrimaryKey(autoGenerate = true)
     var sid: Int? = null,
 
-    @SerializedName("air_date")
+        @SerializedName("air_date")
     @Expose
     var airDate: String? = null,
-    @SerializedName("episode_number")
+        @SerializedName("episode_number")
     @Expose
     var episodeNumber: Int? = null,
-    @SerializedName("id")
+        @SerializedName("id")
     @Expose
     var id: Int? = null,
-    @SerializedName("name")
+        @SerializedName("name")
     @Expose
     var name: String? = null,
-    @SerializedName("overview")
+        @SerializedName("overview")
     @Expose
     var overview: String? = null,
-    @SerializedName("production_code")
+        @SerializedName("production_code")
     @Expose
     var productionCode: String? = null,
-    @SerializedName("season_number")
+        @SerializedName("season_number")
     @Expose
     var seasonNumber: Int? = null,
-    @SerializedName("show_id")
+        @SerializedName("show_id")
     @Expose
     var showId: Int? = null,
-    @SerializedName("still_path")
+        @SerializedName("still_path")
     @Expose
     var stillPath: String? = null,
-    @SerializedName("vote_average")
+        @SerializedName("vote_average")
     @Expose
-    var voteAverage: Int? = null,
-    @SerializedName("vote_count")
+        var voteAverage: Float? = null,
+        @SerializedName("vote_count")
     @Expose
     var voteCount: Int? = null
