+++ app/src/main/java/com/miracozkan/tvseries/datalayer/model/SeriesNextEpisodeToAir.kt
@@ -22,40 +22,40 @@ import kotlinx.android.parcel.RawValue
 @Parcelize
 data class SeriesNextEpisodeToAir(
 
-    @PrimaryKey(autoGenerate = true)
+        @PrimaryKey(autoGenerate = true)
     var sid: Int? = null,
 
-    @SerializedName("air_date")
+        @SerializedName("air_date")
     @Expose
     var airDate: String? = null,
-    @SerializedName("episode_number")
+        @SerializedName("episode_number")
     @Expose
     var episodeNumber: Int? = null,
-    @SerializedName("id")
+        @SerializedName("id")
     @Expose
     var id: Int? = null,
-    @SerializedName("name")
+        @SerializedName("name")
     @Expose
     var name: String? = null,
-    @SerializedName("overview")
+        @SerializedName("overview")
     @Expose
     var overview: String? = null,
-    @SerializedName("production_code")
+        @SerializedName("production_code")
     @Expose
     var productionCode: String? = null,
-    @SerializedName("season_number")
+        @SerializedName("season_number")
     @Expose
     var seasonNumber: Int? = null,
-    @SerializedName("show_id")
+        @SerializedName("show_id")
     @Expose
     var showId: Int? = null,
-    @SerializedName("still_path")
+        @SerializedName("still_path")
     @Expose
     var stillPath: @RawValue Any? = null,
-    @SerializedName("vote_average")
+        @SerializedName("vote_average")
     @Expose
-    var voteAverage: Int? = null,
-    @SerializedName("vote_count")
+        var voteAverage: Float? = null,
+        @SerializedName("vote_count")
     @Expose
     var voteCount: Int? = null
